Refactor[Info]: rename Row component file to InfoRow
