+++ src/lib/features/game/components/Info/InfoRow.tsx
